增加 根据块哈希值读取文件信息功能 增加 选择恢复文件的 ui

+++ BlockInfo.h
@@ -9,7 +9,7 @@
 struct BlockInfo {
     QString filePath;   // 块所在的源文件位置
     qint64 location;    // 块在文件的（起始）位置（在第几Byte）
-    unsigned int size;  // 块的大小（Byte）
+    size_t size;  // 块的大小（Byte）
 };
 
 #endif // BLOCKINFO_H
